+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- New endpoint (`GetResponsesFlatJSON`) and data exporter code to export repsonses in a flat JSON list.
 - Data exporter can parse composite question titles or option labels (when text contains multiple parts).
 - Data exporter logic to handle question type: "responsive single choice array" and "responsive bipolar likert scale array"
 - Participant File Upload endpoint: method to upload files for study participants